+++ src/video/webgl/webgl_renderer.js
@@ -70,6 +70,14 @@ export default class WebGLRenderer extends Renderer {
         this.lineWidth = 1;
 
         /**
+         * sets or returns the shape used to join two line segments where they meet.
+         * Out of the three possible values for this property: "round", "bevel", and "miter", only "round" is supported for now in WebGL
+         * @type {string}
+         * @default "round"
+         */
+        this.lineJoin = "round";
+
+        /**
          * the vertex buffer used by this WebGL Renderer
          * @type {WebGLBuffer}
          */